google-hackathon/ai-research-assistant/requirements.txt
asposestorage==1.0.2
gliner==0.1.12
helpers==0.2.0
langchain==0.1.16
langchain_core==0.1.46
langchain_openai==0.1.4
neo4j==5.20.0
nltk==3.8.1
python-dotenv==1.0.1
unstructured==0.13.6
whyhow==0.0.4
